Remove plugin module
