+++ docs/readme/install.pypi.rst
@@ -9,7 +9,7 @@ so we recommend ensuring setuptools is up to date before installing::
 After awscli-login has been installed, run the following command
 to enable the plugin::
 
-    $ aws configure set plugins.login awscli_login.plugin
+    $ aws configure set plugins.login awscli_login
 
 If you receive a bad interpreter error or other error please see
 the `Known Issues`_ section below.
