+++ docs/readme/readme.rst
@@ -5,8 +5,4 @@ Installation
 
 .. include:: readme/install.{{ ENV }}.rst
 
-Upgrade
-=======
-
-.. include:: readme/upgrade.{{ ENV }}.rst
 .. include:: readme/body.rst
